moved get_bias_field to util. merged implementations of bias field correction
