+++ scripts/correct_raw_data.py
@@ -12,62 +12,11 @@ from joblib import Parallel, delayed, cpu_count
 import math
 import tifffile as tf
 import SimpleITK as sitk
-from util import tqdm_joblib, chunks, imgResample, upload_file_to_s3, S3Url, s3_object_exists
+from util import tqdm_joblib, chunks, imgResample, upload_file_to_s3, S3Url, s3_object_exists, get_bias_field
 
 
 config = Config(connect_timeout=5, retries={'max_attempts': 5})
 
-def correct_bias_field(img, mask=None, scale=1.0, niters=[50, 50, 50, 50]):
-    """Correct bias field in image using the N4ITK algorithm (http://bit.ly/2oFwAun)
-
-    Parameters:
-    ----------
-    img : {SimpleITK.SimpleITK.Image}
-        Input image with bias field.
-    mask : {SimpleITK.SimpleITK.Image}, optional
-        If used, the bias field will only be corrected within the mask. (the default is None, which results in the whole image being corrected.)
-    scale : {float}, optional
-        Scale at which to compute the bias correction. (the default is 0.25, which results in bias correction computed on an image downsampled to 1/4 of it's original size)
-    niters : {list}, optional
-        Number of iterations per resolution. Each additional entry in the list adds an additional resolution at which the bias is estimated. (the default is [50, 50, 50, 50] which results in 50 iterations per resolution at 4 resolutions)
-
-    Returns
-    -------
-    SimpleITK.SimpleITK.Image
-        Bias-corrected image that has the same size and spacing as the input image.
-    """
-
-     # do in case image has 0 intensities
-    # add a small constant that depends on
-    # distribution of intensities in the image
-    stats = sitk.StatisticsImageFilter()
-    stats.Execute(img)
-    std = math.sqrt(stats.GetVariance())
-    img_rescaled = sitk.Cast(img, sitk.sitkFloat32) + 0.1*std
-
-    spacing = np.array(img_rescaled.GetSpacing())/scale
-    img_ds = imgResample(img_rescaled, spacing=spacing)
-    img_ds = sitk.Cast(img_ds, sitk.sitkFloat32)
-
-    # Calculate bias
-    if mask is None:
-        mask = sitk.Image(img_ds.GetSize(), sitk.sitkUInt8)+1
-        mask.CopyInformation(img_ds)
-    else:
-        if type(mask) is not sitk.SimpleITK.Image:
-            mask_sitk = sitk.GetImageFromArray(mask)
-            mask_sitk.CopyInformation(img)
-            mask = mask_sitk
-        mask = imgResample(mask, spacing=spacing)
-
-    img_ds_bc = sitk.N4BiasFieldCorrection(img_ds, mask, 0.001, niters)
-    bias_ds = img_ds_bc / sitk.Cast(img_ds, img_ds_bc.GetPixelID())
-
-    # Upsample bias
-    bias = imgResample(bias_ds, spacing=img.GetSpacing(), size=img.GetSize())
-
-    img_bc = sitk.Cast(img, sitk.sitkFloat32) * sitk.Cast(bias, sitk.sitkFloat32)
-    return img_bc,bias
 
 
 def sum_tiles(files):
@@ -138,7 +87,7 @@ def correct_raw_data(
         sum_tile = sitk.GetImageFromArray(sum_tile)
 
         # get the bias correction tile using N4ITK
-        bias = sitk.GetArrayFromImage(correct_bias_field(sum_tile,scale=1.0)[-1])
+        bias = sitk.GetArrayFromImage(get_bias_field(sum_tile,scale=1.0))
 
         # save bias tile to local directory
         tf.imsave(bias_path, bias)
